requirements.txt
Flask
uwsgi
mysql-connector-python
